split click cli arguments in unit test
When I provide a full string for the shell command, it's always
treated by click as a "unix shell command" (see the docstring of
click.testing.CliRunner.invoke). So the unit tests were failing
on windows because click's CliRunner was calling shlex.split
on a windows path and mangling it, causing it to pass along
a nonexistent input directory.