Fix unit test that uses ITK image
Had to filter out this batch of weird swig warnings because
in our pyproject.toml we configure pytest such that warnings
are treated as errors and this somehow causes a segfault from itk
when any image related functionality is touched. It only happened
on python3.12, not 3.9.

The warnings are things like

```
builtin type SwigPyObject has no __module__ attribute
```