Limit numpy version to avoid dipy incompatbility
numpy 2.4.0 was released on Dec 21, 3 days ago.
dipy has been fixed a while back to be compatible with it:
https://github.com/dipy/dipy/pull/3654
But Dipy has not released since then.
Limiting the numpy version avoids surprises like this in the future,
as this is not the first time we run into a numpy/dipy release
asynchrony issue.
Dependabot will be a nice controlled way to keep them up to date.